cache mechanism on service calls done / poc done

+++ internal/domain/weather.go
@@ -22,25 +22,25 @@ type Condition struct {
 type Current struct {
 	LastUpdatedEpoch int       `json:"last_updated_epoch"`
 	LastUpdated      string    `json:"last_updated"`
-	TempC            int       `json:"temp_c"`
+	TempC            float64   `json:"temp_c"`
 	TempF            float64   `json:"temp_f"`
 	IsDay            int       `json:"is_day"`
 	Condition        Condition `json:"condition"`
 	WindMph          float64   `json:"wind_mph"`
-	WindKph          int       `json:"wind_kph"`
+	WindKph          float64   `json:"wind_kph"`
 	WindDegree       int       `json:"wind_degree"`
 	WindDir          string    `json:"wind_dir"`
-	PressureMb       int       `json:"pressure_mb"`
+	PressureMb       float64   `json:"pressure_mb"`
 	PressureIn       float64   `json:"pressure_in"`
-	PrecipMm         int       `json:"precip_mm"`
-	PrecipIn         int       `json:"precip_in"`
+	PrecipMm         float64   `json:"precip_mm"`
+	PrecipIn         float64   `json:"precip_in"`
 	Humidity         int       `json:"humidity"`
 	Cloud            int       `json:"cloud"`
-	FeelslikeC       int       `json:"feelslike_c"`
-	FeelslikeF       int       `json:"feelslike_f"`
-	VisKm            int       `json:"vis_km"`
-	VisMiles         int       `json:"vis_miles"`
-	Uv               int       `json:"uv"`
+	FeelslikeC       float64   `json:"feelslike_c"`
+	FeelslikeF       float64   `json:"feelslike_f"`
+	VisKm            float64   `json:"vis_km"`
+	VisMiles         float64   `json:"vis_miles"`
+	Uv               float64   `json:"uv"`
 	GustMph          float64   `json:"gust_mph"`
 	GustKph          float64   `json:"gust_kph"`
 }
